+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.28.4
 - dashboard panel editor: fix edit
-
+- dashboard panel img and CSS option 
 
 # 0.28.1
 - fix dashboard id